# Optimize row_constructor Presto function