row_constructor function used to copy data in cases when it was not necessary.
Fix this by using `context->moveOrCopyResult`.

Also, optimize RowVector::copy to copy one column at a time instead of one cell
(value) at a time.

Added a benchmark.

Before:

```
============================================================================
[...]unctions/prestosql/benchmarks/Row.cpp     relative  time/iter   iters/s
============================================================================
noCopy                                                    791.06us     1.26K
copyMostlyFlat                                              4.28ms    233.54
copyMostlyConst                                            24.67ms     40.53
```

After

```
============================================================================
[...]unctions/prestosql/benchmarks/Row.cpp     relative  time/iter   iters/s
============================================================================
noCopy                                                    760.95us     1.31K
copyMostlyFlat                                              1.69ms    592.85
copyMostlyConst                                             1.50ms    664.71
```